Create authentication view model.
