+++ BIM 360 File Transfer/BIM360FileTransfer/Interfaces/IViewModel.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace BIM360FileTransfer.Interfaces
+{
+    public interface IViewModel
+    {
+        string Id { get; }
+        string Name { get; }
+        void TryInvoke(Action action);
+    }
+}
