+++ BIM 360 File Transfer/BIM360FileTransfer/ViewModels/ObserverableObject.cs
@@ -1,0 +1,15 @@
+﻿using System.ComponentModel;
+
+namespace BIM360FileTransfer.ViewModels
+{
+    public abstract class ObserverableObject : INotifyPropertyChanged
+    {
+        public event PropertyChangedEventHandler PropertyChanged;
+
+        protected virtual void OnPropertyChanged(string propertyName = null)
+        {
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+
+        }
+    }
+}
